Moved find_topic_entry() method to TopicsManagementMiddleware
